requirements.txt
gmusicapi==10.1.2
Django==1.11.2
djangorestframework==3.6.3
Flask==0.12.2
Flask-Jsonpify==1.5.0
Flask-SQLAlchemy==2.2
Flask-RESTful==0.3.6
pymongo==3.4.0
